refactor(api): extract pcap analysis engine into capture leaf
Move the PCAP analysis engine out of the internal/api monolith into a new
stdlib+gopacket-only leaf package internal/api/capture (ADR-0006), so the
transport layer keeps only HTTP handlers and Server-coupled file-path policy.

- New package capture (named to avoid colliding with gopacket/.../pcap):
  - analysis DTOs Packet/TimeRange/IPCount/Stats/AnalysisResult (Pcap prefix
    dropped; json tags byte-identical so the wire shape is unchanged)
  - Analyze() engine + layer parsers, structure validation (ValidateStructure),
    and the bounded result Cache (NewCache)
  - snoop.go/snoop_test.go relocated here too: snoop->pcap conversion is pure
    capture-format logic sharing the DLT/magic constants, so it belongs in the
    leaf, not transport (HasSnoopMagic/SnoopToPCAP exported for the upload path)
- De-globalize the analysis cache: the former package-level globalPcapCache is
  now a Server-owned *capture.Cache field, injected in NewServer and the test
  helper (drops the gochecknoglobals exclusion for pcap.go)
- internal/api keeps decodePcapUpload + handlers and the Server file-path-policy
  methods (processInlineData/prepareReplayRequest/validatePcapFilePath/
  tryPcapInDir/resolvePcapAllowedDirs/writeUploadedFile)
- Add api-capture-isolated depguard rule; leaf imports stdlib+gopacket only
- Split the engine across capture.go (types) / cache.go / analyze.go so each
  file stays well under the 600-line ideal

Related to #830
